Day 11, part A solution
Major work was done down an A* path (pun intended) which didn't pan out.
While working down that line of thinking I noticed a mathematical way to
approach this problem, thankfully, since the pathfinding solution was
getting out of hand rather quickly.

Major credit to the author of
http://keekerdc.com/2011/03/hexagon-grids-coordinate-systems-and-distance-calculations/
for helping me understand the hex grid better.