Day 5 part B
Note: I tried this one with a regexp and it was awful, largely due to a
misunderstanding in the problem text: I assumed that the lowercase and
upper case letters needed a 1:1 match before they could be purged, and
thus tried a regex. The regex was matching incorrectly (according to my
misunderstanding) and resulting in the wrong answer.

Once I understood that ALL INSTANCES of the letters (upper and lower)
can be removed, the problem is really quite trivial.