Day 8 part B
I am really quite annoyed with this: I was very happy with my part A
solution (See b63d03d; #58) and it was entirely inappropriate for part B
of the puzzle. Part B made a flat implementation very painful to work
with and as a result I had to tear down almost the entire thing and redo
it from scratch to actually use a tree structure. But it works.