Day 9 part A
Thank you to @rbenua and @msullivan for talking me through it. In an
earlier version I tried to use container/ring to be cute about the
player cycling but ultimately it was just a giant bag of failsauce and
so I went back to the tried and true modular arithmetic. Who knew that
after spending hours debugging my circle pieces that the problem would
be with the simple arithmetic! Special thank you to @rbeunua for very
basic Python implementation that allowed me to verify that my circle
_WAS_ correct(!) and to draw attention to the score tallying.