Day 9 part B
Previous solution took just under 5 hours to compute the wrong answer
(trivially wrong) due to a misreading in the puzzle. My misreading lead
me to track the winning elf from part A on into part B, which is wrong.

This commit adds a solution which computes part A in 0.021 seconds on a
mid 2015 MacBook Air and completes part B in just over a second, at
1.079 seconds, on the same hardware.

"Laughing at the speed of the answer, indeed."