Day 13 part B (#69)
Now with major debugging and a fixed blunder!

Shoutout to https://mk-hill.github.io/TrialAndError/cart-visualizer/ for
helping me isolate the defect in `GetShuttlesByY`:

In 69bfbde2c, line 241:

    241 sort.Slice(ret, func(i, j int) bool { return f.Shuttles[i].X < f.Shuttles[j].X })

This should be:

    241 sort.Slice(ret, func(i, j int) bool { return ret[i].X < ret[j].X })

How this ever worked in part A is a mystery of luck. Notably, there was
causing a problem with a situation like this:

       11111111112222222222
       01234567890123456789
    10 ---+-----+-----+----
    11 ---+-----+----\|
    12 ---+-----+----++-\
    13    |     |    || |
    14 ---+-----+----++-+--
    15 ---+-----+---->^-+--
    16 ---+-----+----++-+--

The shuttle at (15,25) should always move before (15,26) and cause a
collision. However, in some cases the shuttle at (15,26) would move
first and avoid collision. This defect is related to the above defect on
line 241 in 69bfbde2c.